requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-extensions==2.1.2
idna==2.7
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
